drone_py36/requirements_py36.txt
numpy
opencv-python
pyserial
pyrealsense2
torch
torchvision
ultralytics
